requirements.txt
anyio==3.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.3
ecdsa==0.18.0
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jose==1.0.0
pyasn1==0.5.0
pydantic==1.10.7
PyJWT==2.6.0
PyMySQL==1.0.3
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
urllib3==2.0.2
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
